Fix database type mismatch error in product purchase flow
- Fix character varying = integer error in Pesapal callback handler
- Add proper type casting in JOIN conditions between order_items and products tables
- Ensure productId is cast to integer in Order model when creating order items
- Remove unnecessary string casting in order controller queries
- Resolves payment callback failures and ensures smooth purchase flow

+++ server/src/controllers/pesapal.controller.js
@@ -609,7 +609,7 @@ class PesapalController {
           const itemsResult = await pool.query(
             `SELECT oi.*, p.name as product_name, p.seller_id
              FROM order_items oi
-             LEFT JOIN products p ON oi.product_id = p.id
+             LEFT JOIN products p ON oi.product_id::integer = p.id
              WHERE oi.order_id = $1`,
             [orderId]
           );
